resolve projects
