+++ app/src/main/java/com/example/ncast/MainActivity.kt
@@ -15,10 +15,6 @@ class MainActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         _binding = ActivityMainBinding.inflate(layoutInflater)
         setContentView(binding.root)
-        
-        val bottomNavigationView = findViewById<BottomNavigationView>(R.id.bottomNavigation)
-        val navController = findNavController(R.id.nav_host_fragment)
-        bottomNavigationView.setupWithNavController(navController)
        
     }
