+++ app/src/main/java/com/example/ncast/ui/mainApp/MainAppFragment.kt
@@ -1,28 +1,33 @@
-package com.example.ncast.ui.onboarding
+package com.example.ncast.ui.mainApp
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.setupWithNavController
 import com.example.ncast.R
-import com.example.ncast.databinding.FragmentCreateAccountBinding
+import com.example.ncast.databinding.FragmentMainAppBinding
 
-class CreateAccountFragment : Fragment() {
-    private var _binding: FragmentCreateAccountBinding? = null
+class MainAppFragment : Fragment() {
+    private var _binding: FragmentMainAppBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentCreateAccountBinding.inflate(layoutInflater, container, false)
+        _binding = FragmentMainAppBinding.inflate(layoutInflater, container, false)
         return binding.root
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
+        val navHostFragment = childFragmentManager.findFragmentById(R.id.fragmentContainerView2) as NavHostFragment
+        val navController = navHostFragment.navController
+        binding.bottomNavigation.setupWithNavController(navController)
     }
 
     override fun onDestroyView() {
