+++ app/src/main/java/com/example/ncast/ui/mainApp/library/LibraryFragment.kt
@@ -1,4 +1,4 @@
-package com.example.ncast.ui.profile
+package com.example.ncast.ui.mainApp.library
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,13 +7,13 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.ncast.R
 
-class ProfileFragment : Fragment(R.layout.fragment_profile) {
+class LibraryFragment : Fragment(R.layout.fragment_library) {
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.fragment_profile, container, false)
+        return inflater.inflate(R.layout.fragment_library, container, false)
     }
 }
