+++ app/src/main/java/com/example/ncast/ui/mainApp/profile/ProfileFragment.kt
@@ -1,20 +1,19 @@
-package com.example.ncast.ui.home
+package com.example.ncast.ui.mainApp.profile
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.example.ncast.R
 
-class HomeFragment : Fragment(R.layout.fragment_home) {
+class ProfileFragment : Fragment(R.layout.fragment_profile) {
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_home, container, false)
+        return inflater.inflate(R.layout.fragment_profile, container, false)
     }
 }
