+++ app/src/main/java/com/example/ncast/ui/onBoarding/CreateAccountFragment.kt
@@ -1,0 +1,36 @@
+package com.example.ncast.ui.onBoarding
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.navigation.fragment.findNavController
+import com.example.ncast.databinding.FragmentCreateAccountBinding
+
+class CreateAccountFragment : Fragment() {
+    private var _binding: FragmentCreateAccountBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = FragmentCreateAccountBinding.inflate(layoutInflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.signIn.setOnClickListener {
+            findNavController().navigate(CreateAccountFragmentDirections.actionCreateAccountFragmentToMainAppFragment())
+        }
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+}
