+++ app/src/main/java/com/example/ncast/ui/onBoarding/Onboarding2Fragment.kt
@@ -1,21 +1,21 @@
-package com.example.ncast.ui.onboarding
+package com.example.ncast.ui.onBoarding
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.ncast.databinding.FragmentOnboarding3Binding
+import com.example.ncast.databinding.FragmentOnboarding2Binding
 
-class Onboarding3Fragment : Fragment() {
-    private var _binding: FragmentOnboarding3Binding? = null
+class Onboarding2Fragment : Fragment() {
+    private var _binding: FragmentOnboarding2Binding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        _binding = FragmentOnboarding3Binding.inflate(layoutInflater, container, false)
+        _binding = FragmentOnboarding2Binding.inflate(layoutInflater, container, false)
         return binding.root
     }
 
@@ -23,4 +23,5 @@ class Onboarding3Fragment : Fragment() {
         super.onDestroyView()
         _binding = null
     }
+
 }
