+++ app/src/main/res/drawable/bottom_background.xml
@@ -2,7 +2,7 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape="rectangle">
 
-    <solid android:color="@color/lavender"/>
+    <solid android:color="@color/black"/>
     <corners
         android:radius="40dp"/>
