+++ app/src/main/res/drawable/ic_search.xml
