+++ app/src/main/res/drawable/item_selector.xml
@@ -1,7 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
 
-
     <item
         android:state_selected="true"
         android:color="@color/white"/>
