+++ app/src/main/res/layout/fragment_create_account.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.onboarding.CreateAccountFragment">
+    tools:context=".ui.onBoarding.CreateAccountFragment">
 
     <View
         android:layout_width="match_parent"
