+++ app/src/main/res/layout/fragment_library.xml
@@ -4,14 +4,14 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.library.LibraryFragment"
+    tools:context=".ui.mainApp.library.LibraryFragment"
     android:background="@color/white">
 
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Library"
-        android:textColor="@color/lavender"
+        android:textColor="@color/black"
         android:layout_centerInParent="true"
         android:textSize="36sp"
         app:layout_constraintStart_toStartOf="parent"
