+++ app/src/main/res/layout/fragment_onboarding.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.onboarding.OnboardingFragment">
+    tools:context=".ui.onBoarding.OnboardingFragment">
 
     <androidx.viewpager2.widget.ViewPager2
         android:id="@+id/view_pager"
