+++ app/src/main/res/layout/fragment_onboarding1.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.onboarding.Onboarding1Fragment">
+    tools:context=".ui.onBoarding.Onboarding1Fragment">
 
     <View
         android:layout_width="0dp"
