+++ app/src/main/res/layout/fragment_onboarding2.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.onboarding.Onboarding2Fragment">
+    tools:context=".ui.onBoarding.Onboarding2Fragment">
 
     <View
         android:layout_width="0dp"
