+++ app/src/main/res/layout/fragment_onboarding3.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.onboarding.Onboarding3Fragment">
+    tools:context=".ui.onBoarding.Onboarding3Fragment">
 
     <View
         android:layout_width="0dp"
