+++ app/src/main/res/layout/fragment_profile.xml
@@ -5,14 +5,14 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.profile.ProfileFragment"
+    tools:context=".ui.mainApp.profile.ProfileFragment"
     android:background="@color/white">
 
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Profile"
-        android:textColor="@color/lavender"
+        android:textColor="@color/black"
         android:layout_centerInParent="true"
         android:textSize="36sp"
         app:layout_constraintStart_toStartOf="parent"
