+++ app/src/main/res/menu/bottom_menu.xml
@@ -1,21 +1,21 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
     <item
         android:id="@+id/homeFragment"
-        android:icon="@drawable/baseline_home_24"
+        android:icon="@drawable/ic_home"
         android:title="Home" />
 
     <item
         android:id="@+id/searchFragment"
-        android:icon="@drawable/baseline_search_24"
+        android:icon="@drawable/ic_search"
         android:title="Search" />
 
     <item
         android:id="@+id/libraryFragment"
-        android:icon="@drawable/baseline_library_music_24"
+        android:icon="@drawable/ic_library"
         android:title="Library" />
 
     <item
         android:id="@+id/profileFragment"
-        android:icon="@drawable/baseline_person_24"
+        android:icon="@drawable/ic_profile"
         android:title="Profile" />
 </menu>
