+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -7,25 +7,25 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.example.ncast.ui.home.HomeFragment"
+        android:name="com.example.ncast.ui.mainApp.home.HomeFragment"
         android:label="Home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/searchFragment"
-        android:name="com.example.ncast.ui.search.SearchFragment"
+        android:name="com.example.ncast.ui.mainApp.search.SearchFragment"
         android:label="Search"
         tools:layout="@layout/fragment_search" />
 
     <fragment
         android:id="@+id/libraryFragment"
-        android:name="com.example.ncast.ui.library.LibraryFragment"
+        android:name="com.example.ncast.ui.mainApp.library.LibraryFragment"
         android:label="Library"
         tools:layout="@layout/fragment_library" />
 
     <fragment
         android:id="@+id/profileFragment"
-        android:name="com.example.ncast.ui.profile.ProfileFragment"
+        android:name="com.example.ncast.ui.mainApp.profile.ProfileFragment"
         android:label="Profile"
         tools:layout="@layout/fragment_profile" />
 </navigation>
